requirements.txt
# Core
streamlit>=1.32.0
pandas>=2.0.0
requests>=2.28.0

# Semantic search
pinecone-client>=3.0.0
cohere>=5.0.0

# NER pipeline
spacy==3.8.7
numpy==1.26.4
botok==0.9.0

# Tibetan spaCy model (local tarball)
./packages/bo_core_news_lg-0.0.7.tar.gz
